PMD 7.2, fixing api

+++ modules/library/api/src/main/java/org/geotools/api/filter/identity/Version.java
@@ -110,7 +110,7 @@ public final class Version {
             throw new IllegalArgumentException(
                     "Invalid version index: " + index + ". Must be a positive integer > 0.");
         }
-        this.union = UNION_INTEGER | (long) index;
+        this.union = UNION_INTEGER | index;
     }
 
     public Version(final Date dateTime) {
